+++ CopilotKit/packages/react-ui/src/components/chat/Suggestion.tsx
@@ -1,4 +1,4 @@
-import { useCopilotChat } from "@copilotkit/react-core";
+import { useCopilotChatInternal as useCopilotChat } from "@copilotkit/react-core";
 import { SmallSpinnerIcon } from "./Icons";
 
 interface SuggestionsProps {