+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @copilotkit/runtime-client-gql
 
+## 1.10.0-next.7
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.0-next.7
+
+## 1.10.0-next.6
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.0-next.6
+
+## 1.10.0-next.5
+
+### Patch Changes
+
+- Updated dependencies [a8c0263]
+  - @copilotkit/shared@1.10.0-next.5
+
+## 1.10.0-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.0-next.4
+
+## 1.10.0-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.10.0-next.3
+
 ## 1.10.0-next.2
 
 ### Patch Changes